# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title:  Mercado das Mil Sabedorias API
  description: API para conversão de moedas no Reino SRM
  version: 1.0.0
  contact:
    name: teste SRM
    email: api@teste.com
servers:
  - url: 'http://localhost:8080/api/v1'
    description: Servidor de Desenvolvimento

tags:
  - name: Currencies
    description: Operações relacionadas a moedas e taxas de câmbio
  - name: Products
    description: Operações relacionadas a produtos
  - name: Transactions
    description: Operações relacionadas a transações de câmbio

paths:
  /api/v1/currencies/exchange-rate:
    get:
      tags:
        - Currencies
      summary: Obter taxa de câmbio atual entre duas moedas
      operationId: getExchangeRate
      parameters:
        - name: sourceCurrencyCode
          in: query
          description: Código da moeda de origem
          required: true
          schema:
            type: string
        - name: targetCurrencyCode
          in: query
          description: Código da moeda de destino
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Taxa de câmbio obtida com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRate'
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Taxa de câmbio não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Currencies
      summary: AtAPIualizar taxa de câmbio entre duas moedas
      operationId: updateExchangeRate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceCurrencyCode
                - targetCurrencyCode
                - rate
                - effectiveDate
              properties:
                sourceCurrencyCode:
                  type: string
                  description: Código da moeda de origem
                targetCurrencyCode:
                  type: string
                  description: Código da moeda de destino
                rate:
                  type: number
                  format: double
                  description: Taxa de conversão
                effectiveDate:
                  type: string
                  format: date-time
                  description: Data de efetivação da taxa
      responses:
        '200':
          description: Taxa de câmbio atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRate'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products:
    post:
      tags:
        - Products
      summary: Criar um novo produto
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - category
                - kingdom
              properties:
                name:
                  type: string
                  description: Nome do produto
                description:
                  type: string
                  description: Descrição do produto
                category:
                  type: string
                  description: Categoria do produto
                kingdom:
                  type: string
                  description: Reino do produto
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{id}:
    get:
      tags:
        - Products
      summary: Obter produto por ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: ID do produto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Produto obtido com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{id}/conversion-rules:
    post:
      tags:
        - Products
      summary: Criar regra de conversão específica para produto
      operationId: createProductConversionRule
      parameters:
        - name: id
          in: path
          description: ID do produto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - sourceCurrencyCode
                - targetCurrencyCode
                - conversionFormula
                - effectiveDate
              properties:
                productId:
                  type: integer
                  format: int64
                  description: ID do produto
                sourceCurrencyCode:
                  type: string
                  description: Código da moeda de origem
                targetCurrencyCode:
                  type: string
                  description: Código da moeda de destino
                conversionFormula:
                  type: string
                  description: Fórmula de conversão
                effectiveDate:
                  type: string
                  format: date-time
                  description: Data de efetivação da regra
      responses:
        '201':
          description: Regra de conversão criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductConversionRule'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{id}/conversion-rules/current:
    get:
      tags:
        - Products
      summary: Encontrar regra de conversão específica para produto
      operationId: findProductConversionRules
      parameters:
        - name: id
          in: path
          description: ID do produto
          required: true
          schema:
            type: integer
            format: int64
        - name: sourceCurrencyCode
          in: query
          description: Código da moeda de origem
          required: false
          schema:
            type: string
        - name: targetCurrencyCode
          in: query
          description: Código da moeda de destino
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Regras de conversão obtidas com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductConversionRule'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/transactions/convert:
    post:
      tags:
        - Transactions
      summary: Converter moeda para um produto ou conversão geral
      operationId: convertCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - sourceCurrencyCode
                - targetCurrencyCode
              properties:
                productId:
                  type: integer
                  format: int64
                  description: ID do produto (opcional para conversão geral)
                amount:
                  type: number
                  format: double
                  description: Quantidade a ser convertida
                sourceCurrencyCode:
                  type: string
                  description: Código da moeda de origem
                targetCurrencyCode:
                  type: string
                  description: Código da moeda de destino
      responses:
        '201':
          description: Conversão realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Moeda ou produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/transactions:
    get:
      tags:
        - Transactions
      summary: Obter transações com filtros
      operationId: getTransactions
      parameters:
        - name: sourceCurrency
          in: query
          description: Moeda de origem
          required: false
          schema:
            type: string
        - name: targetCurrency
          in: query
          description: Moeda de destino
          required: false
          schema:
            type: string
        - name: kingdom
          in: query
          description: Reino
          required: false
          schema:
            type: string
        - name: fromDate
          in: query
          description: Data inicial
          required: false
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: Data final
          required: false
          schema:
            type: string
            format: date
        - name: status
          in: query
          description: Status da transação
          required: false
          schema:
            type: string
            enum: [PENDING, COMPLETED, FAILED]
        - name: page
          in: query
          description: Número da página
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Tamanho da página
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Transações obtidas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTransaction'
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Currency:
      type: object
      required:
        - code
        - name
        - kingdom
      properties:
        id:
          type: integer
          format: int64
          description: ID da moeda
        code:
          type: string
          description: Código da moeda
        name:
          type: string
          description: Nome da moeda
        kingdom:
          type: string
          description: Reino ao qual pertence a moeda
      example:
        id: 1
        code: "GLD"
        name: "Gold Coin"
        kingdom: "SRM"

    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: double
          description: Valor monetário
        currency:
          $ref: '#/components/schemas/Currency'
      example:
        amount: 100.0
        currency:
          id: 1
          code: "GLD"
          name: "Gold Coin"
          kingdom: "SRM"

    Product:
      type: object
      required:
        - name
        - description
        - category
        - kingdom
      properties:
        id:
          type: integer
          format: int64
          description: ID do produto
        name:
          type: string
          description: Nome do produto
        description:
          type: string
          description: Descrição do produto
        category:
          type: string
          description: Categoria do produto
        kingdom:
          type: string
          description: Reino ao qual pertence o produto
      example:
        id: 1
        name: "Espada Mágica"
        description: "Uma espada lendária forjada nas montanhas de Crystal"
        category: "Armas"
        kingdom: "SRM"

    Transaction:
      type: object
      required:
        - sourceAmount
        - targetAmount
        - appliedRate
        - transactionDate
        - status
      properties:
        id:
          type: integer
          format: int64
          description: ID da transação
        product:
          $ref: '#/components/schemas/Product'
          description: Produto associado à transação (opcional para conversão geral)
        sourceAmount:
          $ref: '#/components/schemas/Money'
          description: Valor na moeda de origem
        targetAmount:
          $ref: '#/components/schemas/Money'
          description: Valor na moeda de destino
        appliedRate:
          type: number
          format: double
          description: Taxa aplicada na conversão
        transactionDate:
          type: string
          format: date-time
          description: Data da transação
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
          description: Status da transação
        notes:
          type: string
          description: Observações sobre a transação
      example:
        id: 1
        product:
          id: 1
          name: "Espada Mágica"
          description: "Uma espada lendária forjada nas montanhas de Crystal"
          category: "Armas"
          kingdom: "SRM"
        sourceAmount:
          amount: 100.0
          currency:
            id: 1
            code: "GLD"
            name: "Gold Coin"
            kingdom: "SRM"
        targetAmount:
          amount: 250.0
          currency:
            id: 2
            code: "SLV"
            name: "Silver Coin"
            kingdom: "SRM"
        appliedRate: 2.5
        transactionDate: "2023-01-01T10:30:00Z"
        status: "COMPLETED"
        notes: "Conversão de ouro para prata para compra de espada mágica"

    ProductConversionRule:
      type: object
      required:
        - product
        - sourceCurrency
        - targetCurrency
        - conversionFormula
        - effectiveDate
        - active
      properties:
        id:
          type: integer
          format: int64
          description: ID da regra de conversão
        product:
          $ref: '#/components/schemas/Product'
          description: Produto associado à regra de conversão
        sourceCurrency:
          $ref: '#/components/schemas/Currency'
          description: Moeda de origem
        targetCurrency:
          $ref: '#/components/schemas/Currency'
          description: Moeda de destino
        conversionFormula:
          type: string
          description: Fórmula de conversão
        effectiveDate:
          type: string
          format: date-time
          description: Data de efetivação da regra
        active:
          type: boolean
          description: Indica se a regra está ativa
      example:
        id: 1
        product:
          id: 1
          name: "Espada Mágica"
          description: "Uma espada lendária forjada nas montanhas de Crystal"
          category: "Armas"
          kingdom: "SRM"
        sourceCurrency:
          id: 1
          code: "GLD"
          name: "Gold Coin"
          kingdom: "SRM"
        targetCurrency:
          id: 2
          code: "SLV"
          name: "Silver Coin"
          kingdom: "SRM"
        conversionFormula: "amount * 2.5"
        effectiveDate: "2023-01-01T00:00:00Z"
        active: true

    ExchangeRate:
      type: object
      required:
        - sourceCurrency
        - targetCurrency
        - rate
        - effectiveDate
      properties:
        id:
          type: integer
          format: int64
          description: ID da taxa de câmbio
        sourceCurrency:
          $ref: '#/components/schemas/Currency'
          description: Moeda de origem
        targetCurrency:
          $ref: '#/components/schemas/Currency'
          description: Moeda de destino
        rate:
          type: number
          format: double
          description: Taxa de conversão
        effectiveDate:
          type: string
          format: date-time
          description: Data de efetivação da taxa
        expirationDate:
          type: string
          format: date-time
          description: Data de expiração da taxa
      example:
        id: 1
        sourceCurrency:
          id: 1
          code: "GLD"
          name: "Gold Coin"
          kingdom: "SRM"
        targetCurrency:
          id: 2
          code: "SLV"
          name: "Silver Coin"
          kingdom: "SRM"
        rate: 2.5
        effectiveDate: "2023-01-01T00:00:00Z"
        expirationDate: "2023-12-31T23:59:59Z"

    PageTransaction:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pageable:
          type: object
          properties:
            pageNumber:
              type: integer
            pageSize:
              type: integer
            sort:
              type: object
              properties:
                sorted:
                  type: boolean
                unsorted:
                  type: boolean
                empty:
                  type: boolean
            offset:
              type: integer
            paged:
              type: boolean
            unpaged:
              type: boolean
        totalPages:
          type: integer
        totalElements:
          type: integer
        last:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        sort:
          type: object
          properties:
            sorted:
              type: boolean
            unsorted:
              type: boolean
            empty:
              type: boolean
        numberOfElements:
          type: integer
        first:
          type: boolean
        empty:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      example:
        timestamp: "2023-01-01T00:00:00Z"
        status: 400
        error: "Bad Request"
        message: "Moeda de origem não encontrada"
        path: "/api/v1/transactions/convert"